refactor(chat): introduce new stream handling utilities and restructure event processing for improved performance and maintainability
